+++ R/gg_bar.R
@@ -178,7 +178,7 @@ gg_bar <- function(data,
     if (is.null(x_expand)) x_expand <- c(0, 0)
     
     if (is.null(x_labels)) {
-      if (is.numeric(x_var_vctr)) x_labels <- scales::label_comma()
+      if (is.numeric(x_var_vctr)) x_labels <- scales::label_number(big.mark = "")
       else if (lubridate::is.Date(x_var_vctr)) x_labels <- scales::label_date()
       else x_labels <- waiver()
     }
@@ -556,7 +556,7 @@ gg_bar_col <- function(data,
     if (is.null(x_expand)) x_expand <- c(0, 0)
     
     if (is.null(x_labels)) {
-      if (is.numeric(x_var_vctr)) x_labels <- scales::label_comma()
+      if (is.numeric(x_var_vctr)) x_labels <- scales::label_number(big.mark = "")
       else if (lubridate::is.Date(x_var_vctr)) x_labels <- scales::label_date()
       else x_labels <- waiver()
     }
@@ -866,7 +866,7 @@ gg_bar_facet <- function(data,
       if (is.null(x_expand)) x_expand <- c(0, 0)
       
       if (is.null(x_labels)) {
-        if (is.numeric(x_var_vctr)) x_labels <- scales::label_comma()
+        if (is.numeric(x_var_vctr)) x_labels <- scales::label_number(big.mark = "")
         else if (lubridate::is.Date(x_var_vctr)) x_labels <- scales::label_date()
         else x_labels <- waiver()
       }
@@ -1251,7 +1251,7 @@ gg_bar_col_facet <- function(data,
       if (is.null(x_expand)) x_expand <- c(0, 0)
       
       if (is.null(x_labels)) {
-        if (is.numeric(x_var_vctr)) x_labels <- scales::label_comma()
+        if (is.numeric(x_var_vctr)) x_labels <- scales::label_number(big.mark = "")
         else if (lubridate::is.Date(x_var_vctr)) x_labels <- scales::label_date()
         else x_labels <- waiver()
       }