+++ R/gg_line.R
@@ -161,7 +161,7 @@ gg_line <- function(data,
     if (is.null(x_expand)) x_expand <- c(0, 0)
     
     if (is.null(x_labels)) {
-      if (is.numeric(x_var_vctr)) x_labels <- scales::label_comma()
+      if (is.numeric(x_var_vctr)) x_labels <- scales::label_number(big.mark = "")
       else if (lubridate::is.Date(x_var_vctr)) x_labels <- scales::label_date()
       else x_labels <- waiver()
     }
@@ -451,7 +451,7 @@ gg_line_col <- function(data,
     if (is.null(x_expand)) x_expand <- c(0, 0)
     
     if (is.null(x_labels)) {
-      if (is.numeric(x_var_vctr)) x_labels <- scales::label_comma()
+      if (is.numeric(x_var_vctr)) x_labels <- scales::label_number(big.mark = "")
       else if (lubridate::is.Date(x_var_vctr)) x_labels <- scales::label_date()
       else x_labels <- waiver()
     }
@@ -742,7 +742,7 @@ gg_line_facet <- function(data,
       if (is.null(x_expand)) x_expand <- c(0, 0)
       
       if (is.null(x_labels)) {
-        if (is.numeric(x_var_vctr)) x_labels <- scales::label_comma()
+        if (is.numeric(x_var_vctr)) x_labels <- scales::label_number(big.mark = "")
         else if (lubridate::is.Date(x_var_vctr)) x_labels <- scales::label_date()
         else x_labels <- waiver()
       }
@@ -1044,7 +1044,7 @@ gg_line_col_facet <- function(data,
       if (is.null(x_expand)) x_expand <- c(0, 0)
       
       if (is.null(x_labels)) {
-        if (is.numeric(x_var_vctr)) x_labels <- scales::label_comma()
+        if (is.numeric(x_var_vctr)) x_labels <- scales::label_number(big.mark = "")
         else if (lubridate::is.Date(x_var_vctr)) x_labels <- scales::label_date()
         else x_labels <- waiver()
       }